+++ CHANGELOG.md
@@ -4,10 +4,10 @@
 
 **Changelog:**
 - Add support for post-quantum ML-KEM from FIPS-203 ([#431](https://github.com/orion-rs/orion/pull/431)).
+- Add support for hybrid KEM X-Wing (draft06 version) ([#434](https://github.com/orion-rs/orion/issues/434)).
 - Implement `core::error::Error` instead of the `std`-version ([#440](https://github.com/orion-rs/orion/pull/440)).
 - Bump MSRV to `1.81.0`.
 - Update CI dependencies.
-
 
 ### 0.17.8